Working Version
In this version I implemented program for passing fixed message between two processes. I understand what is further implementations for the program, but I have to prepare small midterm project for another course, so  I experienced lack of time in proper implementation. Sorry for that. The required table is added to the header of the file.

The problem in my implementation is exchange of send and address variables between processes. I try to implement MPI_Bcast broadcast so that sending process sends to all others index rank of receiving process and the process which just sent the message, but for some reason I did not successed with this idea.