Working version (different task)
In this version I implemented program for passing message with processes' order between two processes. I understand what is further implementations for the program, but I have to prepare small midterm project for another course, so I experienced lack of time in proper implementation. Sorry for that. 

The problem in my implementation is exchange of send and address variables between processes. I try to implement MPI_Bcast broadcast so that sending process sends to all others index rank of receiving process and the process which just sent the message, but for some reason I did not successed with this idea.

Obtained results:
Total time = 15.370568 sec
# of itterations = 500000

Total time = 2.890450 sec
# of itterations = 200000

Total time = 0.956657 sec
# of itterations = 100000